Split pyvm.rs (4205 lines) into focused modules
Break the monolithic pyvm.rs into 5 focused submodules under pyvm/:
- mod.rs: PyVM struct, run/async_run, module registration, tests
- classify.rs: classify_yielded + DoCtrl↔PyObject conversion
- control_primitives.rs: All pyclass control types + nesting helpers
- helpers.rs: Error conversion, type normalization, metadata extraction
- run_result.rs: PyDoeffTracebackData + PyRunResult

All 30 Rust tests pass. No functional changes.

Closes gh-295
